% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doinrtrocrop.R
\name{do.anova}
\alias{do.anova}
\title{this fucnction is for ANOVA on two species.}
\usage{
do.anova(InputPath = NULL, inputdf = NULL, alphavalue = NULL)
}
\arguments{
\item{InputPath}{is path of input csv file.}

\item{inputdf}{is a dataframe}

\item{alphavalue}{value of alpha for LSD.test and duncan
this file has six columns that are Zab,Zba,B,R,SYa,SYb
Zab is  ratio of first species  to second species; Zba vise versa.
B is value Ratio.
R is Replication.
SYa is Yield of first species.
SYb is Yield of second species.
It is noteworthy, mensioned description is for all function of input path}
}
\description{
this fucnction is for ANOVA on two species.
}
\examples{
do.anova(InputPath="d:\\\\input.csv")
}
