% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doinrtrocrop.R
\name{do.anova.sunflower}
\alias{do.anova.sunflower}
\title{this fucnction is for ANOVA on sunflower}
\usage{
do.anova.sunflower(inputDataPath)
}
\arguments{
\item{inputDataPath}{is path for csv input file
this file has six column that are Zbs,Zsb,B,R,SYb,SYs
Zbs is for example ratio of Beans to sunflower Zsb vise versa
B is value Ratio and R is Replication
in that example Yb is Yield of Beans
and Ys is Yield of Sunflower
It is noteworthy, mensioned description is for all function of input path}
}
\description{
this fucnction is for ANOVA on sunflower
}
\examples{
do.anova.sunflower("d:\\\\input.csv")
}
