% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doinrtrocrop.R
\name{do.intercropping.indies}
\alias{do.intercropping.indies}
\title{this function is for intercropping and  calculate total relative value, namely Aggressivity, Actual yield loss, relative yield, RVT and ...}
\usage{
do.intercropping.indies(
  InputPath = NULL,
  OutputPath,
  inputdf = NULL,
  PriceOfFirstSpeies = NULL,
  PriceOfSecondSpeies = NULL
)
}
\arguments{
\item{InputPath}{is path of input csv file.
this file has six columns that are Zab,Zba,B,R,SYa,SYb
Zab is  ratio of first species  to second species; Zba vise versa.
B is value Ratio.
R is Replication.
SYa is Yield of first species.
SYb is Yield of second species.
It is noteworthy, mensioned description is for all function of input path}

\item{OutputPath}{is path of output of total relative value}

\item{inputdf}{is a dataframe for input data}

\item{PriceOfFirstSpeies}{is a number namely  is price of First Speies}

\item{PriceOfSecondSpeies}{is a number namely  is price of Second Speies}
}
\value{
return a dataframe
}
\description{
this function is for intercropping and  calculate total relative value, namely Aggressivity, Actual yield loss, relative yield, RVT and ...
}
\examples{
do.intercropping.indies(InputPath="d:\\\\input.csv",OutputPath="d:\\\\output.csv")
}
