% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doinrtrocrop.R
\name{do.otherindices.intercrop}
\alias{do.otherindices.intercrop}
\title{this function is for calculate other indices namely Aggressivity, Actual yield loss, relative yield, RVT and ...}
\usage{
do.otherindices.intercrop(
  beanDatPath,
  outputAddress,
  priceofbean = NULL,
  priceofsunflower = NULL
)
}
\arguments{
\item{beanDatPath}{path for input csv file
this file has six column that are Zbs,Zsb,B,R,SYb,SYs
Zbs is for example ratio of Beans to sunflower Zsb vise versa
B is value Ratio and R is Replication
in that example Yb is Yield of Beans
and Ys is Yield of Sunflower
It is noteworthy, mensioned description is for all function of input path}

\item{outputAddress}{path and file name for output csv file}

\item{priceofbean}{is number and is price of bean}

\item{priceofsunflower}{is number and is price of sunflower}
}
\value{
return a dataframe
}
\description{
this function is for calculate other indices namely Aggressivity, Actual yield loss, relative yield, RVT and ...
}
\examples{
do.otherindices.intercrop("d:\\\\input.csv","d:\\\\otherindicesoutput.csv")
do.otherindices.intercrop("d:\\\\input.csv","d:\\\\otherindicesoutput.csv",100000,20000)
}
