% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doinrtrocrop.R
\name{do.summary.intercrop}
\alias{do.summary.intercrop}
\title{is a function for summary output of intercropping}
\usage{
do.summary.intercrop(beanDatPath, outputAddress)
}
\arguments{
\item{beanDatPath}{is path of input csv file.
this file has six column that are Zbs,Zsb,B,R,SYb,SYs
Zbs is for example ratio of Beans to sunflower Zsb vise versa
B is value Ratio and R is Replication
in that example Yb is Yield of Beans
and Ys is Yield of Sunflower
It is noteworthy, mensioned description is for all function of input path}

\item{outputAddress}{is path of summery output of intercropping}
}
\description{
is a function for summary output of intercropping
}
\examples{
do.summary.intercrop("d:\\\\input.csv","d:\\\\output.csv")
}
