% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputdata.R
\docType{data}
\name{inputdata}
\alias{inputdata}
\title{is a dataframe for test}
\format{
inputdata is a dataframe
}
\usage{
inputdata
}
\description{
use this dataframe for doing intercropping between two species
}
